requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
confluent-kafka
stripe
python-decouple
APScheduler